requirements.txt
# Core dependencies
gradio>=4.0.0
python-dotenv>=1.0.0

# Google Gemini API
google-genai>=0.3.0

# Image processing
Pillow>=10.0.0

# Text-to-Speech engines
gTTS>=2.5.0
ChatTTS>=0.1.0

# ChatTTS dependencies
transformers==4.53.2
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.0

# Additional utilities
numpy>=1.24.0
